Refactor: create session.py for session
Fix: import session from xonsh.session

Refactor: don't pass XSH.ctx before it is initialised
